+++ sdk/core/azure-core/inc/azure/match_conditions.hpp
@@ -1,0 +1,33 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+/**
+ * @file
+ * @brief Define MatchConditions
+ */
+
+#pragma once
+
+#include "azure/etag.hpp"
+
+#include <string>
+
+namespace Azure {
+
+/**
+ * @brief Specifies HTTP options for conditional requests.
+ */
+struct MatchConditions
+{
+  /**
+   * @brief Optionally limit requests to resources that match the value specified.
+   */
+  ETag IfMatch;
+
+  /**
+   * @brief Optionally limit requests to resources that do not match the value specified. Specify
+   * Azure::ETag::Any() to limit requests to resources that do not exist.
+   */
+  ETag IfNoneMatch;
+};
+} // namespace Azure