+++ sdk/core/azure-core/test/ut/match_conditions.cpp
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/core/match_conditions.hpp>
 #include <azure/etag.hpp>
+#include <azure/match_conditions.hpp>
 #include <gtest/gtest.h>
 #include <string>