+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -39,8 +39,8 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::Http::RawResponse r(
       1, 1, Azure::Core::Http::HttpStatusCode::Accepted, "phrase"));
   EXPECT_NO_THROW(Azure::Core::Logger::Listener ll = nullptr);
-  EXPECT_NO_THROW(Azure::Core::MatchConditions mc);
-  EXPECT_NO_THROW(Azure::Core::ModifiedConditions mc);
+  EXPECT_NO_THROW(Azure::MatchConditions mc);
+  EXPECT_NO_THROW(Azure::ModifiedConditions mc);
   EXPECT_NO_THROW(Azure::Core::Nullable<int> n);
   EXPECT_NO_THROW(Azure::Core::Uuid::CreateUuid());
   EXPECT_NO_THROW(Azure::Core::RequestFailedException("foo"));